refactor:파일 크기 타입 일관성 유지

+++ src/main/java/com/back/b2st/domain/performance/controller/AdminPerformanceController.java
@@ -127,7 +127,10 @@ public class AdminPerformanceController {
 	public BaseResponse<PresignedUrlRes> generatePresignedUrl(
 		@Valid @RequestBody CreatePresignedUrlReq request
 	) {
-		return BaseResponse.success(performanceService.generatePosterPresign(request.contentType(), request.fileSize().longValue()));
+		return BaseResponse.success(performanceService.generatePosterPresign(
+			request.contentType(),
+			request.fileSize()
+		));
 	}
 }
